+++ adv_optm/optim/Lion_Prodigy_adv.py
@@ -225,8 +225,8 @@ class Lion_Prodigy_adv(torch.optim.Optimizer):
                 # Pre-generate random tensor for stochastic rounding if needed.
                 random_int_tensor = param_update._get_random_int_for_sr(p)
             # TODO, workaround until pytorch#169634 is fixed
-            d = torch.as_tensor(group['d'])
-            dlr = torch.as_tensor(dlr)
+            d = torch.as_tensor(group['d'], dtype=torch.float64)
+            dlr = torch.as_tensor(dlr, dtype=torch.float64)
             step_param_fn = self._compiled_step_parameter
         else:
             d = group['d']
